Fix aarch64 package
The noarch package with __linux dependency pulled in the {{ compiler() }} for
linux-64, since it was built on that platform.

We don't have a generic, platform-independent `c-compiler` package which is
resolved at time of package use. So we can't make this a noarch package.

Bumping the build number in addition, and removing the abs.yaml configuration
addressing noarch package upload.